More efficient search for the root_device
We used to calculate everything twice, once for the "base" planned
volumes (only / and the boot-related ones) just to decide which disk to
use as root_device and then again with all the planned volumes.

But there was probably no real gain in doing that and it was confusing to
trace and debug.