fix miscalculation of sizes when resizing a Windows partition (bsc#1066386)
What otherwise happens is this: our planning strategy will typically resize
an existing Windows partition *twice*. First for the root partition, then
again for the rest.

Without this patch, after the first (correct) resize, the Windows partition
will leave a free space block that appears to be bigger than can actually be
used (due to the alignment issue). As a result, the second resize would be
planned too small.