rely on storage ids instead of device names in SpaceMaker#provide_space
This also rewrites PartitionKiller to use storage ids.

The space maker basically collects the storage ids for partition names when
it starts.

To fully switch to using storage ids instead of names everywhere in the
proposal calculation is not feasible as our unit test mockup is not prepared
for this.