allow for numbers > 32 bit in region dialog (bsc#1065258)
libyui's IntField widget works only with 32 bits. Since the major visible
advantage over a plain InputField are the controls allowing to
increase/decrease the value by one - which is not very helpful here - I
think it's fine to use an InputField here.

I've also changed the logic for the min/max values presented: it used to
show the first and last available blocks regardless whether these belong to
the same unused region.  This runs into the problem that in nearly all cases
you run into an error if you just confirm the defaults (for instance the min
value used to be _before_ the first partition).

Instead it now shows the default (largest) region (which is always a valid
range).