Improve type label for Ext3/4 devices with journal
The Ext3/4 multi device filesystems only use two devices: one for data
and other for the external journal. Thus, it does not make sense to show
"Part of ext3/4 (/whatever)" as type for the data device. Instead, seems
better to display "Ext3/4 Partition" for it and "Journal for ext3/4"
(/whatever)" for the external journal.