Restore the .label pages method
Sometime ago[1][2], the .label method from some pages was moved to the
UIState in order to avoid a dependency cycles. However, it was a kind of
temporary fix and it's time to really solve the problem.

As first step, this commit "restore" the .label method for those pages.

Subsequent commit(s) will avoid mentioned dependency.

[1] https://github.com/yast/yast-storage-ng/pull/752
[2] https://github.com/yast/yast-storage-ng/pull/752/commits/b107fc5b13662b10fe8904f7d6331a2b057eaa61