Bring back limit to speed up initial proposal (bsc#1154070)
The limit of 5 disks was introduced in the SLE-15-GA branch to solve
the mentioned bug in system with many disks. In SLE-15-SP1 it was
considered to not be necessary because YaST always tries individual
disks first. Now that ProposalSettings#multidisk_first has been introduced,
the limit becomes again relevant, so this commit reintroduces it.