update /etc/crypttab status properly (bsc#1118977)
The #update_etc_status method is called when you switch between mounting/not
mounting in the fstab dialog.

But when encryption had just been enabled, the luks device does not yet
exist in the device tree. It's created only after going through the password
dialog.

So the #update_etc_status method has to be called also after creating an
encryption device.