Do not remove a Btrfs subvolume for updating it
For letting to change the path of Btrfs subvolumes in memory (those that
does not exist on disk yet), the partitioner was deleting and creating
it again. Since it is a wrong approach to perform such operation, from
now on none kind of Btrfs subvolume is going to be deleted when updating
it.