Discard the StorageManager#staging! method
Because we agreed that it's safe (and somehow better) to change the
StorageManager#staging for using #probe! instead of #probe since the
generic Yast::AbortException is properly handled in WFM#run_client[1].
This avoids having to check and change every call to #staging in the
rest of YaST modules and/or clients.

Partially invalidates commit 7f41574a6708cbfd222d9b6629f792f3efd18bda

[1] https://github.com/yast/yast-ruby-bindings/blob/c5de99becfd9b154e995f4a615b8748923c3a088/src/ruby/yast/wfm.rb#L358