[Proposal] Pre-calculate the list of planned devices
This is the second step to reorganize a bit the GuidedProposal algorithm
to fix some issues and to allow new use cases. It divides the calculation
of planned devices in two phases, one before cleaning the devicegraph
(mandatory actions) and a second phase afterwards (to mark reusable swaps
and to calculate boot-related planned partitions).